Remove action hook when saving stellar settings
The original idea was to send a request to Stellar when the new stellar settings were saved and then show an admin notice if
the account did not have the correct flags set. The admin_init hook is triggered before this, meaning the stellar_destination_tag_check function was being called before this meaning the values set in this hooks were being ignored until next reload.

The new approach is to run this function being only on the stellar settings page and in certain situations i.e ( new account_address, the check again button was clicked in the notice and etc.)